Redid Matrix, Vector classes
I redid large parts of the Matrix and Vector classes. I added a few new operators, restructured stuff, and further tweaked the code (most of my changes were things like making an argument a const or using unsigned in the place of signed ints and stuff like that).

The code compiles, but has not been tested.