Tweaks to Gauss Seidel + Vector print method
I added a print method to the vector class. This basically just prints out all of the contents of the vector.

I also modified the iterativeLA code. Most of the changes that I made were minor syntax changes. The code appears to work correctly.

I plan on adding SOR code in the next commit.