Added SOR + Updated error processing
I added a SOR function to  my code. This function is pretty similar to my Gauss_Seidel code. I also added in an extra check to make sure that the relaxation paramter is appropiate.

To make it easier to remember how to use the various fucntions, I added some comments to the jacobi and Gauss_Siedel prototypes (in iterativeLA.h). I did the same for the new SOR function. See iterativeLA.h.

I also added a new case to the switch statement (in main.cc) to handle the state "BAD_DATA". I also modified the Jacobi/GS/SOR code so that the various checks, if triggered, print out some information about which check was tripped and why (see itertiveLA.cc)

This code compiles and has been tested.