Added Method choice + Diagional dominant checker
I  modified the main.cc code in two major ways. First, I added code to check if the entered matrix, A, is diagionally dominant. I wrote a function (in the Matrix files) to check for strict diagional dominace. See Matrix.cc:195.

I also modified the main code so that the user is prompted to choose a method (Jacobi, Gauss Siedel, or SOR). This was done using an enumerator (to give the different methods proper labels). See Main.cc:50 for detials.

I also changed jacobi to Jacobi and addes in some more spaces (see iterativeLA.cc)

this code compiles and runs. It has been tested.